+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/codestyle/LocalVariableNamingConventionsRule.java
@@ -38,6 +38,10 @@ public final class LocalVariableNamingConventionsRule extends AbstractNamingConv
 
     @Override
     public Object visit(ASTVariableId node, Object data) {
+        if (node.isUnnamed()) {
+            // unnamed variables do not have to match the regexes.
+            return null;
+        }
 
         if (node.isExceptionBlockParameter()) {
             checkMatches(node, exceptionBlockParameterRegex, data);